# clarityhubbackend/docs/openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clarity Store API",
    "version": "1.0.0",
    "description": "Comprehensive E-commerce API Documentation for Clarity Store",
    "contact": {
      "name": "Clarity Support",
      "email": "support@claritystore.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Access token stored in HttpOnly cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/orders/my-orders": {
      "get": {
        "summary": "Get user orders",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user orders."
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get all orders (Admin)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all orders."
          }
        }
      }
    },
    "/orders/export": {
      "get": {
        "summary": "Export orders (Admin)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Downloadable CSV/Excel report."
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order details (Admin)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Order details."
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "summary": "Update order status (Admin)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "processing",
                      "shipped",
                      "delivered",
                      "cancelled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated."
          }
        }
      }
    },
...
    "/orders/{id}/refund": {
      "post": {
        "summary": "Process refund (Admin)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund processed."
          }
        }
      }
    },
...
    "/products": {
      "post": {
        "summary": "Create a new product (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "price",
                  "category",
                  "stock"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "category": {
                    "type": "string"
                  },
                  "stock": {
                    "type": "number"
                  },
                  "image": {
                    "type": "string",
                    "description": "Base64 image string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully."
          },
          "403": {
            "description": "Unauthorized - Admin role required."
          }
        }
      },
      "get": {
        "summary": "Get all products (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all products."
          }
        }
      }
    },
...
    "/products/search": {
      "get": {
        "summary": "Search products",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search query string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Search results."
          }
        }
      }
    },
    "/products/category/{category}": {
      "get": {
        "summary": "Get products by category",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Products in the specified category."
          }
        }
      }
    },
    "/products/slug/{slug}": {
      "get": {
        "summary": "Get product by slug",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product details."
          },
          "404": {
            "description": "Product not found."
          }
        }
      }
    },
    "/products/low-stock": {
      "get": {
        "summary": "Get low stock products (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of items with low stock."
          }
        }
      }
    },
    "/products/bulk": {
      "patch": {
        "summary": "Bulk update products (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products updated."
          }
        }
      },
      "delete": {
        "summary": "Bulk delete products (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products deleted."
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get product details (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product details."
          }
        }
      },
      "put": {
        "summary": "Update product (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated."
          }
        }
      },
      "delete": {
        "summary": "Delete product (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted."
          }
        }
      }
    },
    "/products/{id}/featured": {
      "patch": {
        "summary": "Toggle product featured status (Admin)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Status toggled."
          }
        }
      }
    },
...
  },
...
}